requirements.txt
sounddevice
numpy
scipy
openai-whisper
requests
tk
ffmpeg-python
ollama

